added common test cases for new expense managament sw page
