+++ expense-management-software/test_expense_management_software.py
@@ -2,16 +2,28 @@ from time import sleep
 import logging
 import pytest
 from common.utils import resize_browser
-from common.asserts import assert_spacing_between, assert_spacing_bottom, assert_spacing_top, assert_spacing_right, assert_spacing_left
+from common.asserts import assert_spacing_between, assert_spacing_bottom, assert_spacing_top, assert_spacing_right, assert_spacing_left, assert_overflowing, assert_customer_testimonial, assert_customer_logo, assert_typography
 
 logger = logging.getLogger(__name__)
 
 @pytest.fixture(scope='function')
 def browser(module_browser, base_url, request):
     resize_browser(browser=module_browser, resolution=request.param)
-    module_browser.get(base_url + '/travel-and-expense-software')
+    module_browser.get(base_url + '/expense-management-software')
     sleep(4)
     return module_browser
+
+@pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
+def test_overflowing(browser):
+    assert_overflowing(browser=browser)
+
+@pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
+def test_typography(browser):
+    assert_typography(browser=browser)
+
+@pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
+def test_customer_logo(browser):
+    assert_customer_logo(browser=browser)
 
 @pytest.mark.parametrize('browser', [('desktop_1')], indirect=True)
 def test_competitor_section_spacing(browser):
